+++ alicloud/resource_alicloud_vswitch_test.go
@@ -106,7 +106,7 @@ func testSweepVSwitches(region string) error {
 	return nil
 }
 
-func TestAccAlicloudVSwitch_basic(t *testing.T) {
+func TestAccAlicloudVSwitch_Update(t *testing.T) {
 	var vsw vpc.DescribeVSwitchAttributesResponse
 
 	resource.Test(t, resource.TestCase{
@@ -123,13 +123,40 @@ func TestAccAlicloudVSwitch_basic(t *testing.T) {
 				Config: testAccVswitchConfig,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckVswitchExists("alicloud_vswitch.foo", &vsw),
-					resource.TestCheckResourceAttr(
-						"alicloud_vswitch.foo", "cidr_block", "172.16.0.0/21"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "cidr_block", "172.16.0.0/21"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "name", "tf-testAccVswitchConfig"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "description", ""),
+				),
+			},
+			{
+				Config: testAccVswitchConfigRename,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckVswitchExists("alicloud_vswitch.foo", &vsw),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "cidr_block", "172.16.0.0/21"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "name", "tf-testAccVswitchConfigRename"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "description", ""),
+				),
+			},
+			{
+				Config: testAccVswitchConfigRedesc,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckVswitchExists("alicloud_vswitch.foo", &vsw),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "cidr_block", "172.16.0.0/21"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "name", "tf-testAccVswitchConfigRename"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "description", "I am MrX"),
+				),
+			},
+			{
+				Config: testAccVswitchConfigUpdate,
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckVswitchExists("alicloud_vswitch.foo", &vsw),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "cidr_block", "172.16.0.0/21"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "name", "tf-testAccVswitchConfigUpdate"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo", "description", "How Are You"),
 				),
 			},
 		},
 	})
-
 }
 
 func TestAccAlicloudVSwitch_multi(t *testing.T) {
@@ -148,14 +175,11 @@ func TestAccAlicloudVSwitch_multi(t *testing.T) {
 				Config: testAccVswitchMulti,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckVswitchExists("alicloud_vswitch.foo_0", &vsw),
-					resource.TestCheckResourceAttr(
-						"alicloud_vswitch.foo_0", "cidr_block", "172.16.0.0/24"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo_0", "cidr_block", "172.16.0.0/24"),
 					testAccCheckVswitchExists("alicloud_vswitch.foo_1", &vsw),
-					resource.TestCheckResourceAttr(
-						"alicloud_vswitch.foo_1", "cidr_block", "172.16.1.0/24"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo_1", "cidr_block", "172.16.1.0/24"),
 					testAccCheckVswitchExists("alicloud_vswitch.foo_2", &vsw),
-					resource.TestCheckResourceAttr(
-						"alicloud_vswitch.foo_2", "cidr_block", "172.16.2.0/24"),
+					resource.TestCheckResourceAttr("alicloud_vswitch.foo_2", "cidr_block", "172.16.2.0/24"),
 				),
 			},
 		},
@@ -227,6 +251,65 @@ resource "alicloud_vswitch" "foo" {
   cidr_block = "172.16.0.0/21"
   availability_zone = "${data.alicloud_zones.default.zones.0.id}"
   name = "${var.name}"
+}
+`
+const testAccVswitchConfigRename = `
+data "alicloud_zones" "default" {
+	"available_resource_creation"= "VSwitch"
+}
+variable "name" {
+  default = "tf-testAccVswitchConfigRename"
+}
+resource "alicloud_vpc" "foo" {
+  name = "${var.name}"
+  cidr_block = "172.16.0.0/12"
+}
+
+resource "alicloud_vswitch" "foo" {
+  vpc_id = "${alicloud_vpc.foo.id}"
+  cidr_block = "172.16.0.0/21"
+  availability_zone = "${data.alicloud_zones.default.zones.0.id}"
+  name = "${var.name}"
+}
+`
+const testAccVswitchConfigRedesc = `
+data "alicloud_zones" "default" {
+	"available_resource_creation"= "VSwitch"
+}
+variable "name" {
+  default = "tf-testAccVswitchConfigRename"
+}
+resource "alicloud_vpc" "foo" {
+  name = "${var.name}"
+  cidr_block = "172.16.0.0/12"
+}
+
+resource "alicloud_vswitch" "foo" {
+  vpc_id = "${alicloud_vpc.foo.id}"
+  cidr_block = "172.16.0.0/21"
+  availability_zone = "${data.alicloud_zones.default.zones.0.id}"
+  name = "${var.name}"
+  description="I am MrX"
+}
+`
+const testAccVswitchConfigUpdate = `
+data "alicloud_zones" "default" {
+	"available_resource_creation"= "VSwitch"
+}
+variable "name" {
+  default = "tf-testAccVswitchConfigUpdate"
+}
+resource "alicloud_vpc" "foo" {
+  name = "${var.name}"
+  cidr_block = "172.16.0.0/12"
+}
+
+resource "alicloud_vswitch" "foo" {
+  vpc_id = "${alicloud_vpc.foo.id}"
+  cidr_block = "172.16.0.0/21"
+  availability_zone = "${data.alicloud_zones.default.zones.0.id}"
+  name = "${var.name}"
+  description="How Are You"
 }
 `