+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 IMPROVEMENTS:
 
+- Apply wraperror to ess group [GH-689]
+- Add wraperror and apply it to vpc and eip [GH-688]
+- Improve vswitch resource and data source testcases [GH-687]
+- Improve security_group resource and data source testcases [GH-686]
+- Improve vpc resource and data source testcases [GH-684]
 - Modify the slb sever group testcase name [GH-681]
 - Improve sweeper testcases [GH-680]
 - Improve db instance's testcases [GH-679]